requirements.txt
fastapi
uvicorn
langchain
chromadb
sentence-transformers
httpx
python-dotenv
streamlit
PyPDF2
